Alerts

+++ src/app/components/character-list/character-list.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
+import { MatDialog } from '@angular/material';
 import { Character } from '../../shared/models/character.model';
 import { CharacterService } from '../../shared/service/character.service';
-import { AuthServiceService } from '../user/auth-service.service';
+import { UserService } from '../user/user.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 /**
  * Character-list component gets a list of character traits from the database.
@@ -14,22 +16,36 @@ import { AuthServiceService } from '../user/auth-service.service';
 })
 export class CharacterListComponent implements OnInit, OnDestroy {
   /**
-   * characters property used to reference an array of character data.
+   * @property characters used to reference an array of character data.
    */
   characters: Character[] = [];
   /**
-   * characterSub property with a type of Subscription from rxjs library.
+   * @property characterSub rxjs Subscription
    * Unsubscribes in ngOnDestroy function.
    */
   private characterSub: Subscription;
+  /**
+   * @property userId string
+   */
   userId: string;
+  /**
+   * Checks a user's authentication status.
+   */
   userIsAuthenticated = false;
+  /**
+   * @property authStatusSub rxjs Subscription
+   */
   private authStatusSub: Subscription;
+  title = 'confirmation-dialog';
 
   /**
    *  @ignore
    */
-  constructor(public characterService: CharacterService, private authService: AuthServiceService) { }
+  constructor(
+    public characterService: CharacterService,
+    private userService: UserService,
+    public dialog: MatDialog
+    ) { }
 
   /**
    * ngOnInit function performs a GET request for list of character traits
@@ -37,25 +53,34 @@ export class CharacterListComponent implements OnInit, OnDestroy {
    */
   ngOnInit() {
     this.characterService.getCharacters();
-    this.userId = this.authService.getUserId();
+    this.userId = this.userService.getUserId();
     this.characterSub = this.characterService.getCharacterUpdateListener()
       .subscribe((characters: Character[]) => {
         this.characters = characters;
       });
-    this.userIsAuthenticated = this.authService.getIsAuth();
-    this.authStatusSub = this.authService
+    this.userIsAuthenticated = this.userService.getIsAuth();
+    this.authStatusSub = this.userService
       .getAuthStatusListener()
       .subscribe(isAuthenticated => {
         this.userIsAuthenticated = isAuthenticated;
-        this.userId = this.authService.getUserId();
+        this.userId = this.userService.getUserId();
       });
   }
+
   /**
-   * Performs a delete function from the characterService on a button click.
+   * Opens a dialog popup when the delete button is clicked
    * @param characterId string
    */
-  onDelete(characterId: string) {
-    this.characterService.deleteCharacter(characterId);
+  openDialog(characterId: string) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '350px',
+      data: 'Are you sure you want this deleted?'
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.characterService.deleteCharacter(characterId);
+      }
+    });
   }
 
   /**
